RealdashCAN Serial Send Protocol.
Implementation of the RealdashCAN serial protocol on Speeduino serial 03. I am implementing the RealdashCAN Serial protocol on Speeduino Serial 03 with Mega 2560. -The code takes all the data and sends it every 8 bytes with a specific ID and sends it at a frequency of 200Hz. -Connection to Realdash is done as RealdashCAN. -Only some are mapped in XML, for now. -I’m planning to implement CAN Broadcast for the STM32 platform, so the same XML can be used. Anyone who can do testing will be of great help.